Ensure surreal decorators call underlying route hook when completed
Previously, the hooks defined in the @ascua/surreal route decorators would completely override any hooks defined in the consuming route. Now, the overridden hook is stored, and once the @ascua/surreal route hook is completed, the underlying route hook is called.

In addition, we no longer await for the surreal connection to be established and authentication to be attempted, as this embedded promise causes problems in Ember FastBoot. Instead, the application route should be marked as an @attempted route, causing all other beforeModel() hooks to be called only once authenication has been attempted.