Wait for session to be identified before proceeding
When a route was redirected after signup or signin, the route’s model hook would fire before the session was ready. If there model hook required a value from the session model, then the route would fail. Now, we track when the session model has been loaded, and when the session is ready to be used, before proceeding to an authenticated route.