Fix further TransitionAborted errors in Fastboot
When running in Fastboot,aborting a transition inside an asynchronous promise, still causes an error when in the redirect hook, unlike is stated here: https://github.com/ember-fastboot/ember-cli-fastboot/issues/202. Instead we will not wait for any Surreal connection attempt, instead relying on the user to use the @attempted decorator in the application route.