Rearrange order of image attributes
In Safari, as the attributes are changed in the order that they are specified by glimmer, the src would be the first attribute to be set once the fastboot hydration had begun. This caused the browser to download the image in the src attribute even though it was not needed. Now, the sizes attribute is specified first, ensuring that the sizes are known by the browser before the srcset, and finally, the src attributes are specified.