Ensure the ‘_super’ property is not enumerable on Arrays
Currently Ember.js has a bug which causes the ‘_super’ property to be enumerable on Arrays, when `EXTEND_PROTOTYPES` is enabled. This causes a bug in PDF.js ES5 version: “The `Array.prototype` contains unexpected enumerable properties: _super; thus breaking e.g. `for...in` iteration of `Array`s.”

This temporary fix ensures that the ‘_super’ property is not enumerable, until the fix can be made to Ember.js. See [this issue](https://github.com/emberjs/ember.js/issues/19289) for more information.